ai3d/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
google-generativeai>=0.6.0
python-multipart
